test_hasfulltablecache
